Final polish: split tests/support/mocks.rs into per-endpoint files
Moves MockModelEndpoint, MockOpenAIEndpoint, and MockChatEndpoint into
their own files under mocks/. HTTP helpers and request predicates stay
in mocks/mod.rs. SSE helpers (tool_call_sse, completion_text_sse) live
in mocks/chat.rs and are re-exported from mod.rs for backwards compat.
