--- crates/defra-agent-cli/tests/support/mocks.rs
@@ -1,677 +1,0 @@
-use std::io::{Read, Write};
-use std::net::{Shutdown, TcpListener, TcpStream};
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Arc, Mutex};
-use std::thread;
-use std::thread::JoinHandle;
-use std::time::Duration;
-
-use anyhow::{anyhow, Context, Result};
-use serde_json::Value;
-
-pub struct MockModelEndpoint {
-    pub endpoint: String,
-    pub port: u16,
-    pub stop: Arc<AtomicBool>,
-    pub handle: Option<JoinHandle<()>>,
-}
-
-pub struct MockOpenAIEndpoint {
-    pub endpoint: String,
-    pub port: u16,
-    pub stop: Arc<AtomicBool>,
-    pub captured_chat_requests: Arc<Mutex<Vec<Value>>>,
-    pub handle: Option<JoinHandle<()>>,
-}
-
-pub struct MockChatEndpoint {
-    pub endpoint: String,
-    pub port: u16,
-    pub stop: Arc<AtomicBool>,
-    pub captured_chat_requests: Arc<Mutex<Vec<Value>>>,
-    pub handle: Option<JoinHandle<()>>,
-}
-
-pub struct HttpRequestData {
-    pub method: String,
-    pub path: String,
-    pub headers: std::collections::HashMap<String, String>,
-    pub body: Vec<u8>,
-}
-
-impl MockModelEndpoint {
-    pub fn start(model_name: &str) -> Result<Self> {
-        Self::start_with_required_bearer(model_name, None)
-    }
-
-    pub fn start_with_required_bearer(model_name: &str, required_bearer: Option<&str>) -> Result<Self> {
-        let listener = TcpListener::bind(("127.0.0.1", 0)).context("binding mock model port")?;
-        listener
-            .set_nonblocking(true)
-            .context("marking mock model listener nonblocking")?;
-        let port = listener
-            .local_addr()
-            .context("reading mock model port")?
-            .port();
-        let stop = Arc::new(AtomicBool::new(false));
-        let stop_for_thread = stop.clone();
-        let model_name = model_name.to_string();
-        let required_bearer = required_bearer.map(ToOwned::to_owned);
-        let handle = thread::spawn(move || {
-            while !stop_for_thread.load(Ordering::Relaxed) {
-                match listener.accept() {
-                    Ok((mut stream, _)) => {
-                        let request = match read_http_request(&mut stream) {
-                            Ok(request) => request,
-                            Err(_) => {
-                                let _ = stream.shutdown(Shutdown::Both);
-                                continue;
-                            }
-                        };
-
-                        let authorized = required_bearer.as_ref().is_none_or(|expected| {
-                            request
-                                .headers
-                                .get("authorization")
-                                .is_some_and(|value| value == &format!("Bearer {expected}"))
-                        });
-                        let (status, body) = if request.method == "GET"
-                            && (request.path == "/v1/models" || request.path == "/models")
-                        {
-                            if authorized {
-                                ("200 OK", format!(r#"{{"data":[{{"id":"{model_name}"}}]}}"#))
-                            } else {
-                                (
-                                    "401 Unauthorized",
-                                    r#"{"error":"unauthorized"}"#.to_string(),
-                                )
-                            }
-                        } else {
-                            ("404 Not Found", r#"{"error":"not found"}"#.to_string())
-                        };
-                        let _ = write_http_response(&mut stream, status, "application/json", &body);
-                        let _ = stream.shutdown(Shutdown::Both);
-                    }
-                    Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
-                        thread::sleep(Duration::from_millis(25));
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        Ok(Self {
-            endpoint: format!("http://127.0.0.1:{port}/v1"),
-            port,
-            stop,
-            handle: Some(handle),
-        })
-    }
-
-    pub fn endpoint(&self) -> &str {
-        &self.endpoint
-    }
-}
-
-impl Drop for MockModelEndpoint {
-    fn drop(&mut self) {
-        self.stop.store(true, Ordering::Relaxed);
-        let _ = TcpStream::connect(("127.0.0.1", self.port));
-        if let Some(handle) = self.handle.take() {
-            let _ = handle.join();
-        }
-    }
-}
-
-impl MockOpenAIEndpoint {
-    pub fn start(model_name: &str, final_token: &str) -> Result<Self> {
-        let listener = TcpListener::bind(("127.0.0.1", 0)).context("binding mock OpenAI port")?;
-        listener
-            .set_nonblocking(true)
-            .context("marking mock OpenAI listener nonblocking")?;
-        let port = listener
-            .local_addr()
-            .context("reading mock OpenAI port")?
-            .port();
-        let stop = Arc::new(AtomicBool::new(false));
-        let stop_for_thread = stop.clone();
-        let model_name = model_name.to_string();
-        let final_token = final_token.to_string();
-        let captured_chat_requests = Arc::new(Mutex::new(Vec::new()));
-        let captured_chat_requests_for_thread = captured_chat_requests.clone();
-
-        let handle = thread::spawn(move || {
-            while !stop_for_thread.load(Ordering::Relaxed) {
-                match listener.accept() {
-                    Ok((mut stream, _)) => {
-                        let request = match read_http_request(&mut stream) {
-                            Ok(request) => request,
-                            Err(_) => {
-                                let _ = stream.shutdown(Shutdown::Both);
-                                continue;
-                            }
-                        };
-
-                        match (request.method.as_str(), request.path.as_str()) {
-                            ("GET", "/v1/models") => {
-                                let body = format!(r#"{{"data":[{{"id":"{model_name}"}}]}}"#);
-                                let _ = write_http_response(
-                                    &mut stream,
-                                    "200 OK",
-                                    "application/json",
-                                    &body,
-                                );
-                            }
-                            ("POST", "/v1/chat/completions") => {
-                                let request_json: Value =
-                                    match serde_json::from_slice(&request.body) {
-                                        Ok(value) => value,
-                                        Err(_) => {
-                                            let _ = write_http_response(
-                                                &mut stream,
-                                                "400 Bad Request",
-                                                "application/json",
-                                                r#"{"error":"invalid json"}"#,
-                                            );
-                                            let _ = stream.shutdown(Shutdown::Both);
-                                            continue;
-                                        }
-                                    };
-                                captured_chat_requests_for_thread
-                                    .lock()
-                                    .expect("captured chat request mutex poisoned")
-                                    .push(request_json.clone());
-
-                                let sse_body = if request_has_tool_result_message(&request_json) {
-                                    completion_text_sse(&final_token)
-                                } else {
-                                    tool_call_sse("read_file", r#"{"path":"notes.txt"}"#)
-                                };
-                                let _ = write_http_response(
-                                    &mut stream,
-                                    "200 OK",
-                                    "text/event-stream",
-                                    &sse_body,
-                                );
-                            }
-                            _ => {
-                                let _ = write_http_response(
-                                    &mut stream,
-                                    "404 Not Found",
-                                    "application/json",
-                                    r#"{"error":"not found"}"#,
-                                );
-                            }
-                        }
-
-                        let _ = stream.flush();
-                        let _ = stream.shutdown(Shutdown::Both);
-                    }
-                    Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
-                        thread::sleep(Duration::from_millis(25));
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        Ok(Self {
-            endpoint: format!("http://127.0.0.1:{port}/v1"),
-            port,
-            stop,
-            captured_chat_requests,
-            handle: Some(handle),
-        })
-    }
-
-    pub fn endpoint(&self) -> &str {
-        &self.endpoint
-    }
-
-    pub fn captured_chat_requests(&self) -> Vec<Value> {
-        self.captured_chat_requests
-            .lock()
-            .expect("captured chat request mutex poisoned")
-            .clone()
-    }
-}
-
-impl Drop for MockOpenAIEndpoint {
-    fn drop(&mut self) {
-        self.stop.store(true, Ordering::Relaxed);
-        let _ = TcpStream::connect(("127.0.0.1", self.port));
-        if let Some(handle) = self.handle.take() {
-            let _ = handle.join();
-        }
-    }
-}
-
-impl MockChatEndpoint {
-    pub fn start(model_name: &str, final_text: &str) -> Result<Self> {
-        Self::start_with_required_bearer(model_name, final_text, None)
-    }
-
-    pub fn start_with_required_bearer(
-        model_name: &str,
-        final_text: &str,
-        required_bearer: Option<&str>,
-    ) -> Result<Self> {
-        let listener = TcpListener::bind(("127.0.0.1", 0)).context("binding mock chat port")?;
-        listener
-            .set_nonblocking(true)
-            .context("marking mock chat listener nonblocking")?;
-        let port = listener
-            .local_addr()
-            .context("reading mock chat port")?
-            .port();
-        let stop = Arc::new(AtomicBool::new(false));
-        let stop_for_thread = stop.clone();
-        let model_name = model_name.to_string();
-        let final_text = final_text.to_string();
-        let required_bearer = required_bearer.map(ToOwned::to_owned);
-        let captured_chat_requests = Arc::new(Mutex::new(Vec::new()));
-        let captured_chat_requests_for_thread = captured_chat_requests.clone();
-
-        let handle = thread::spawn(move || {
-            while !stop_for_thread.load(Ordering::Relaxed) {
-                match listener.accept() {
-                    Ok((mut stream, _)) => {
-                        let request = match read_http_request(&mut stream) {
-                            Ok(request) => request,
-                            Err(_) => {
-                                let _ = stream.shutdown(Shutdown::Both);
-                                continue;
-                            }
-                        };
-
-                        let authorized = required_bearer.as_ref().is_none_or(|expected| {
-                            request
-                                .headers
-                                .get("authorization")
-                                .is_some_and(|value| value == &format!("Bearer {expected}"))
-                        });
-
-                        match (request.method.as_str(), request.path.as_str()) {
-                            ("GET", "/v1/models") => {
-                                let (status, body) = if authorized {
-                                    ("200 OK", format!(r#"{{"data":[{{"id":"{model_name}"}}]}}"#))
-                                } else {
-                                    (
-                                        "401 Unauthorized",
-                                        r#"{"error":"unauthorized"}"#.to_string(),
-                                    )
-                                };
-                                let _ = write_http_response(
-                                    &mut stream,
-                                    status,
-                                    "application/json",
-                                    &body,
-                                );
-                            }
-                            ("POST", "/v1/chat/completions") => {
-                                if !authorized {
-                                    let _ = write_http_response(
-                                        &mut stream,
-                                        "401 Unauthorized",
-                                        "application/json",
-                                        r#"{"error":"unauthorized"}"#,
-                                    );
-                                    let _ = stream.shutdown(Shutdown::Both);
-                                    continue;
-                                }
-                                let request_json: Value =
-                                    match serde_json::from_slice(&request.body) {
-                                        Ok(value) => value,
-                                        Err(_) => {
-                                            let _ = write_http_response(
-                                                &mut stream,
-                                                "400 Bad Request",
-                                                "application/json",
-                                                r#"{"error":"invalid json"}"#,
-                                            );
-                                            let _ = stream.shutdown(Shutdown::Both);
-                                            continue;
-                                        }
-                                    };
-                                captured_chat_requests_for_thread
-                                    .lock()
-                                    .expect("captured chat request mutex poisoned")
-                                    .push(request_json);
-
-                                let _ = write_http_response(
-                                    &mut stream,
-                                    "200 OK",
-                                    "text/event-stream",
-                                    &completion_text_sse(&final_text),
-                                );
-                            }
-                            _ => {
-                                let _ = write_http_response(
-                                    &mut stream,
-                                    "404 Not Found",
-                                    "application/json",
-                                    r#"{"error":"not found"}"#,
-                                );
-                            }
-                        }
-
-                        let _ = stream.flush();
-                        let _ = stream.shutdown(Shutdown::Both);
-                    }
-                    Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
-                        thread::sleep(Duration::from_millis(25));
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-
-        Ok(Self {
-            endpoint: format!("http://127.0.0.1:{port}/v1"),
-            port,
-            stop,
-            captured_chat_requests,
-            handle: Some(handle),
-        })
-    }
-
-    pub fn endpoint(&self) -> &str {
-        &self.endpoint
-    }
-
-    pub fn captured_chat_requests(&self) -> Vec<Value> {
-        self.captured_chat_requests
-            .lock()
-            .expect("captured chat request mutex poisoned")
-            .clone()
-    }
-}
-
-impl Drop for MockChatEndpoint {
-    fn drop(&mut self) {
-        self.stop.store(true, Ordering::Relaxed);
-        let _ = TcpStream::connect(("127.0.0.1", self.port));
-        if let Some(handle) = self.handle.take() {
-            let _ = handle.join();
-        }
-    }
-}
-
-pub fn read_http_request(stream: &mut TcpStream) -> Result<HttpRequestData> {
-    let mut buffer = Vec::new();
-    let mut chunk = [0_u8; 4096];
-    let mut header_end = None;
-    let mut content_length = 0_usize;
-
-    loop {
-        let read = stream.read(&mut chunk).context("reading mock request")?;
-        if read == 0 {
-            break;
-        }
-        buffer.extend_from_slice(&chunk[..read]);
-
-        if header_end.is_none() {
-            if let Some(offset) = find_header_end(&buffer) {
-                let end = offset + 4;
-                let headers = String::from_utf8_lossy(&buffer[..end]);
-                header_end = Some(end);
-                content_length = parse_content_length(&headers).unwrap_or(0);
-                if buffer.len() >= end + content_length {
-                    break;
-                }
-            }
-        } else if buffer.len() >= header_end.expect("header_end should be set") + content_length {
-            break;
-        }
-    }
-
-    let header_end = header_end.ok_or_else(|| anyhow!("missing request headers"))?;
-    let header_text = String::from_utf8_lossy(&buffer[..header_end]);
-    let request_line = header_text
-        .lines()
-        .next()
-        .ok_or_else(|| anyhow!("missing request line"))?;
-    let mut parts = request_line.split_whitespace();
-    let method = parts
-        .next()
-        .ok_or_else(|| anyhow!("missing HTTP method"))?
-        .to_string();
-    let path = parts
-        .next()
-        .ok_or_else(|| anyhow!("missing HTTP path"))?
-        .to_string();
-    let headers = header_text
-        .lines()
-        .skip(1)
-        .filter_map(|line| {
-            let (name, value) = line.split_once(':')?;
-            Some((name.trim().to_ascii_lowercase(), value.trim().to_string()))
-        })
-        .collect();
-    let body_end = header_end + content_length;
-    let body = if buffer.len() >= body_end {
-        buffer[header_end..body_end].to_vec()
-    } else {
-        Vec::new()
-    };
-
-    Ok(HttpRequestData {
-        method,
-        path,
-        headers,
-        body,
-    })
-}
-
-pub fn find_header_end(buffer: &[u8]) -> Option<usize> {
-    buffer.windows(4).position(|window| window == b"\r\n\r\n")
-}
-
-pub fn parse_content_length(headers: &str) -> Option<usize> {
-    headers.lines().find_map(|line| {
-        let (name, value) = line.split_once(':')?;
-        if name.eq_ignore_ascii_case("content-length") {
-            value.trim().parse().ok()
-        } else {
-            None
-        }
-    })
-}
-
-pub fn write_http_response(
-    stream: &mut TcpStream,
-    status: &str,
-    content_type: &str,
-    body: &str,
-) -> Result<()> {
-    let response = format!(
-        "HTTP/1.1 {status}\r\ncontent-type: {content_type}\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{body}",
-        body.len()
-    );
-    stream
-        .write_all(response.as_bytes())
-        .context("writing mock response")?;
-    Ok(())
-}
-
-pub fn tool_call_sse(tool_name: &str, arguments: &str) -> String {
-    let chunk_1 = serde_json::json!({
-        "choices": [{
-            "delta": {
-                "tool_calls": [{
-                    "index": 0,
-                    "id": "call-read-file",
-                    "function": {
-                        "name": tool_name,
-                        "arguments": ""
-                    }
-                }]
-            },
-            "finish_reason": null
-        }],
-        "usage": null
-    });
-    let chunk_2 = serde_json::json!({
-        "choices": [{
-            "delta": {
-                "tool_calls": [{
-                    "index": 0,
-                    "id": null,
-                    "function": {
-                        "name": null,
-                        "arguments": arguments
-                    }
-                }]
-            },
-            "finish_reason": null
-        }],
-        "usage": null
-    });
-    let chunk_3 = serde_json::json!({
-        "choices": [{
-            "delta": {
-                "content": null,
-                "tool_calls": []
-            },
-            "finish_reason": "tool_calls"
-        }],
-        "usage": {
-            "prompt_tokens": 16,
-            "completion_tokens": 4,
-            "total_tokens": 20
-        }
-    });
-    format!(
-        "data: {}\n\ndata: {}\n\ndata: {}\n\ndata: [DONE]\n\n",
-        serde_json::to_string(&chunk_1).expect("serialize tool-call chunk 1"),
-        serde_json::to_string(&chunk_2).expect("serialize tool-call chunk 2"),
-        serde_json::to_string(&chunk_3).expect("serialize tool-call chunk 3"),
-    )
-}
-
-pub fn completion_text_sse(text: &str) -> String {
-    let chunk_1 = serde_json::json!({
-        "choices": [{
-            "delta": {
-                "content": text
-            },
-            "finish_reason": null
-        }],
-        "usage": null
-    });
-    let chunk_2 = serde_json::json!({
-        "choices": [{
-            "delta": {
-                "content": null,
-                "tool_calls": []
-            },
-            "finish_reason": "stop"
-        }],
-        "usage": {
-            "prompt_tokens": 24,
-            "completion_tokens": 6,
-            "total_tokens": 30
-        }
-    });
-    format!(
-        "data: {}\n\ndata: {}\n\ndata: [DONE]\n\n",
-        serde_json::to_string(&chunk_1).expect("serialize completion chunk 1"),
-        serde_json::to_string(&chunk_2).expect("serialize completion chunk 2"),
-    )
-}
-
-pub fn request_has_tool_result_message(request: &Value) -> bool {
-    request
-        .get("messages")
-        .and_then(Value::as_array)
-        .is_some_and(|messages| {
-            messages
-                .iter()
-                .any(|message| message.get("role").and_then(Value::as_str) == Some("tool"))
-        })
-}
-
-pub fn request_tool_names(request: &Value) -> Vec<String> {
-    request
-        .get("tools")
-        .and_then(Value::as_array)
-        .map(|tools| {
-            tools
-                .iter()
-                .filter_map(|tool| {
-                    tool.get("function")
-                        .and_then(|function| function.get("name"))
-                        .and_then(Value::as_str)
-                        .map(ToString::to_string)
-                })
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default()
-}
-
-pub fn request_system_message(request: &Value) -> Option<&str> {
-    request
-        .get("messages")
-        .and_then(Value::as_array)
-        .and_then(|messages| {
-            messages.iter().find_map(|message| {
-                if message.get("role").and_then(Value::as_str) != Some("system") {
-                    return None;
-                }
-                match message.get("content") {
-                    Some(Value::String(content)) => Some(content.as_str()),
-                    Some(Value::Array(parts)) => parts
-                        .iter()
-                        .find_map(|part| part.get("text").and_then(Value::as_str)),
-                    _ => None,
-                }
-            })
-        })
-}
-
-pub fn request_tool_result_text(request: &Value) -> Option<String> {
-    request
-        .get("messages")
-        .and_then(Value::as_array)
-        .and_then(|messages| {
-            messages.iter().find_map(|message| {
-                if message.get("role").and_then(Value::as_str) != Some("tool") {
-                    return None;
-                }
-                match message.get("content") {
-                    Some(Value::String(content)) => Some(content.to_string()),
-                    Some(Value::Array(parts)) => {
-                        let text = parts
-                            .iter()
-                            .filter_map(|part| part.get("text").and_then(Value::as_str))
-                            .collect::<Vec<_>>()
-                            .join("\n");
-                        Some(text)
-                    }
-                    _ => None,
-                }
-            })
-        })
-}
-
-pub fn request_contains_role_text(request: &Value, role: &str, needle: &str) -> bool {
-    request
-        .get("messages")
-        .and_then(Value::as_array)
-        .is_some_and(|messages| {
-            messages.iter().any(|message| {
-                if message.get("role").and_then(Value::as_str) != Some(role) {
-                    return false;
-                }
-                match message.get("content") {
-                    Some(Value::String(content)) => content.contains(needle),
-                    Some(Value::Array(parts)) => parts.iter().any(|part| {
-                        part.get("text")
-                            .and_then(Value::as_str)
-                            .is_some_and(|text| text.contains(needle))
-                    }),
-                    _ => false,
-                }
-            })
-        })
-}
