+++ crates/defra-agent-cli/tests/support/mocks/chat.rs
@@ -1,0 +1,257 @@
+use std::io::Write;
+use std::net::{Shutdown, TcpListener, TcpStream};
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::{Arc, Mutex};
+use std::thread;
+use std::thread::JoinHandle;
+use std::time::Duration;
+
+use anyhow::{Context, Result};
+use serde_json::Value;
+
+use super::{read_http_request, write_http_response};
+
+pub struct MockChatEndpoint {
+    pub endpoint: String,
+    pub port: u16,
+    pub stop: Arc<AtomicBool>,
+    pub captured_chat_requests: Arc<Mutex<Vec<Value>>>,
+    pub handle: Option<JoinHandle<()>>,
+}
+
+impl MockChatEndpoint {
+    pub fn start(model_name: &str, final_text: &str) -> Result<Self> {
+        Self::start_with_required_bearer(model_name, final_text, None)
+    }
+
+    pub fn start_with_required_bearer(
+        model_name: &str,
+        final_text: &str,
+        required_bearer: Option<&str>,
+    ) -> Result<Self> {
+        let listener = TcpListener::bind(("127.0.0.1", 0)).context("binding mock chat port")?;
+        listener
+            .set_nonblocking(true)
+            .context("marking mock chat listener nonblocking")?;
+        let port = listener
+            .local_addr()
+            .context("reading mock chat port")?
+            .port();
+        let stop = Arc::new(AtomicBool::new(false));
+        let stop_for_thread = stop.clone();
+        let model_name = model_name.to_string();
+        let final_text = final_text.to_string();
+        let required_bearer = required_bearer.map(ToOwned::to_owned);
+        let captured_chat_requests = Arc::new(Mutex::new(Vec::new()));
+        let captured_chat_requests_for_thread = captured_chat_requests.clone();
+
+        let handle = thread::spawn(move || {
+            while !stop_for_thread.load(Ordering::Relaxed) {
+                match listener.accept() {
+                    Ok((mut stream, _)) => {
+                        let request = match read_http_request(&mut stream) {
+                            Ok(request) => request,
+                            Err(_) => {
+                                let _ = stream.shutdown(Shutdown::Both);
+                                continue;
+                            }
+                        };
+
+                        let authorized = required_bearer.as_ref().is_none_or(|expected| {
+                            request
+                                .headers
+                                .get("authorization")
+                                .is_some_and(|value| value == &format!("Bearer {expected}"))
+                        });
+
+                        match (request.method.as_str(), request.path.as_str()) {
+                            ("GET", "/v1/models") => {
+                                let (status, body) = if authorized {
+                                    ("200 OK", format!(r#"{{"data":[{{"id":"{model_name}"}}]}}"#))
+                                } else {
+                                    (
+                                        "401 Unauthorized",
+                                        r#"{"error":"unauthorized"}"#.to_string(),
+                                    )
+                                };
+                                let _ = write_http_response(
+                                    &mut stream,
+                                    status,
+                                    "application/json",
+                                    &body,
+                                );
+                            }
+                            ("POST", "/v1/chat/completions") => {
+                                if !authorized {
+                                    let _ = write_http_response(
+                                        &mut stream,
+                                        "401 Unauthorized",
+                                        "application/json",
+                                        r#"{"error":"unauthorized"}"#,
+                                    );
+                                    let _ = stream.shutdown(Shutdown::Both);
+                                    continue;
+                                }
+                                let request_json: Value =
+                                    match serde_json::from_slice(&request.body) {
+                                        Ok(value) => value,
+                                        Err(_) => {
+                                            let _ = write_http_response(
+                                                &mut stream,
+                                                "400 Bad Request",
+                                                "application/json",
+                                                r#"{"error":"invalid json"}"#,
+                                            );
+                                            let _ = stream.shutdown(Shutdown::Both);
+                                            continue;
+                                        }
+                                    };
+                                captured_chat_requests_for_thread
+                                    .lock()
+                                    .expect("captured chat request mutex poisoned")
+                                    .push(request_json);
+
+                                let _ = write_http_response(
+                                    &mut stream,
+                                    "200 OK",
+                                    "text/event-stream",
+                                    &completion_text_sse(&final_text),
+                                );
+                            }
+                            _ => {
+                                let _ = write_http_response(
+                                    &mut stream,
+                                    "404 Not Found",
+                                    "application/json",
+                                    r#"{"error":"not found"}"#,
+                                );
+                            }
+                        }
+
+                        let _ = stream.flush();
+                        let _ = stream.shutdown(Shutdown::Both);
+                    }
+                    Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
+                        thread::sleep(Duration::from_millis(25));
+                    }
+                    Err(_) => break,
+                }
+            }
+        });
+
+        Ok(Self {
+            endpoint: format!("http://127.0.0.1:{port}/v1"),
+            port,
+            stop,
+            captured_chat_requests,
+            handle: Some(handle),
+        })
+    }
+
+    pub fn endpoint(&self) -> &str {
+        &self.endpoint
+    }
+
+    pub fn captured_chat_requests(&self) -> Vec<Value> {
+        self.captured_chat_requests
+            .lock()
+            .expect("captured chat request mutex poisoned")
+            .clone()
+    }
+}
+
+impl Drop for MockChatEndpoint {
+    fn drop(&mut self) {
+        self.stop.store(true, Ordering::Relaxed);
+        let _ = TcpStream::connect(("127.0.0.1", self.port));
+        if let Some(handle) = self.handle.take() {
+            let _ = handle.join();
+        }
+    }
+}
+
+pub fn tool_call_sse(tool_name: &str, arguments: &str) -> String {
+    let chunk_1 = serde_json::json!({
+        "choices": [{
+            "delta": {
+                "tool_calls": [{
+                    "index": 0,
+                    "id": "call-read-file",
+                    "function": {
+                        "name": tool_name,
+                        "arguments": ""
+                    }
+                }]
+            },
+            "finish_reason": null
+        }],
+        "usage": null
+    });
+    let chunk_2 = serde_json::json!({
+        "choices": [{
+            "delta": {
+                "tool_calls": [{
+                    "index": 0,
+                    "id": null,
+                    "function": {
+                        "name": null,
+                        "arguments": arguments
+                    }
+                }]
+            },
+            "finish_reason": null
+        }],
+        "usage": null
+    });
+    let chunk_3 = serde_json::json!({
+        "choices": [{
+            "delta": {
+                "content": null,
+                "tool_calls": []
+            },
+            "finish_reason": "tool_calls"
+        }],
+        "usage": {
+            "prompt_tokens": 16,
+            "completion_tokens": 4,
+            "total_tokens": 20
+        }
+    });
+    format!(
+        "data: {}\n\ndata: {}\n\ndata: {}\n\ndata: [DONE]\n\n",
+        serde_json::to_string(&chunk_1).expect("serialize tool-call chunk 1"),
+        serde_json::to_string(&chunk_2).expect("serialize tool-call chunk 2"),
+        serde_json::to_string(&chunk_3).expect("serialize tool-call chunk 3"),
+    )
+}
+
+pub fn completion_text_sse(text: &str) -> String {
+    let chunk_1 = serde_json::json!({
+        "choices": [{
+            "delta": {
+                "content": text
+            },
+            "finish_reason": null
+        }],
+        "usage": null
+    });
+    let chunk_2 = serde_json::json!({
+        "choices": [{
+            "delta": {
+                "content": null,
+                "tool_calls": []
+            },
+            "finish_reason": "stop"
+        }],
+        "usage": {
+            "prompt_tokens": 24,
+            "completion_tokens": 6,
+            "total_tokens": 30
+        }
+    });
+    format!(
+        "data: {}\n\ndata: {}\n\ndata: [DONE]\n\n",
+        serde_json::to_string(&chunk_1).expect("serialize completion chunk 1"),
+        serde_json::to_string(&chunk_2).expect("serialize completion chunk 2"),
+    )
+}
